fix bug and add unit test case for runtime api when shared heap is enabled

+++ core/iwasm/common/wasm_memory.h
@@ -84,7 +84,7 @@ SET_LINEAR_MEMORY_SIZE(WASMMemoryInstance *memory, uint64 size)
 #if WASM_ENABLE_SHARED_HEAP != 0
 bool
 is_app_addr_in_shared_heap(WASMModuleInstanceCommon *module_inst,
-                           bool is_memory64, uint64 app_offset, uint32 bytes);
+                           bool is_memory64, uint64 app_offset, uint64 bytes);
 
 WASMSharedHeap *
 wasm_runtime_create_shared_heap(SharedHeapInitArgs *init_args);
